added validation errors bind to slot
